Version 2 Beta
Major changes. Upgraded to the latest version of the Svbtle CSS and
Javascript. Added the anchor for permalinks. Redid kudos Javascript
actions by modifying the Svbtle JS. Added licensing information to
README. Note that this is the BETA version of 2.0 so there may be some
bugs. Currently in testing at http://erictendian.tk/.